+++ contracts/TestBokkyPooBahsRedBlackTreeRaw.sol
@@ -19,37 +19,37 @@ contract TestBokkyPooBahsRedBlackTreeRaw {
 
     BokkyPooBahsRedBlackTreeLibrary.Tree tree;
 
-    event Log(string where, uint key, uint value);
+    event Log(string where, uint80 key, uint value);
 
     constructor() public {
     }
-    function root() public view returns (uint _key) {
+    function root() public view returns (uint80 _key) {
         _key = tree.root;
     }
-    function first() public view returns (uint _key) {
+    function first() public view returns (uint80 _key) {
         _key = tree.first();
     }
-    function last() public view returns (uint _key) {
+    function last() public view returns (uint80 _key) {
         _key = tree.last();
     }
-    function next(uint key) public view returns (uint _key) {
+    function next(uint80 key) public view returns (uint80 _key) {
         _key = tree.next(key);
     }
-    function prev(uint key) public view returns (uint _key) {
+    function prev(uint80 key) public view returns (uint80 _key) {
         _key = tree.prev(key);
     }
-    function exists(uint key) public view returns (bool _exists) {
+    function exists(uint80 key) public view returns (bool _exists) {
         _exists = tree.exists(key);
     }
-    function getNode(uint _key) public view returns (uint key, uint parent, uint left, uint right, bool red) {
+    function getNode(uint80 _key) public view returns (uint80 key, uint80 parent, uint80 left, uint80 right, bool red) {
         (key, parent, left, right, red) = tree.getNode(_key);
     }
 
-    function insert(uint _key) public {
+    function insert(uint80 _key) public {
         tree.insert(_key);
         // emit Log("insert", _key, 0);
     }
-    function remove(uint _key) public {
+    function remove(uint80 _key) public {
         tree.remove(_key);
         // emit Log("remove", _key, 0);
     }
